add explicit test checks in reset

+++ test/utils/reset.ts
@@ -6,14 +6,14 @@ import { Pool } from "pg";
 import { Logger } from "../../src/utils/logger";
 
 export async function resetRedis() {
-    if (config?.redis?.enabled) {
+    if (config?.redis?.enabled && config.mode === "test") {
         const client = createClient(config.redis);
         await client.connect();
         await client.flushAll();
     }
 }
 export async function resetPostgres() {
-    if (process.env.TEST_POSTGRES && config.postgres) {
+    if (process.env.TEST_POSTGRES && config.mode == "test" && config.postgres) {
         const pool = new Pool({ ...config.postgres });
         await pool.query(`DROP DATABASE IF EXISTS "sponsorTimes"`);
         await pool.query(`DROP DATABASE IF EXISTS "privateDB"`);
